chore(pragma): consolidate MCP tools into domain-owned specs
Move MCP tool definitions from the monolithic mcp/tools/ directory into
each domain as declarative ToolSpec arrays. Domains export specs with
zero MCP imports; the mcp/ adapter layer consumes them via
registerFromSpec() which generates zod schemas and wires server.registerTool().

Architecture:
- New ToolSpec type in domains/shared/ToolSpec.ts (no MCP deps)
- Each domain exports ToolSpec[] from mcp/specs.ts
- registerFromSpec.ts is the single adapter (ToolSpec → MCP)
- Filter helpers moved to domains/shared/filters/
- domains/ never imports from mcp/ (correct dependency direction)

Closes F.13 ST.01 from session.remediation.
