+++ packages/cli/pragma/src/domains/block/index.ts
@@ -12,4 +12,5 @@ export function commands(ctx: PragmaContext): CommandDefinition[] {
   return [listCommand(ctx), lookupCommand(ctx)];
 }
 
+export { specs as mcpSpecs } from "./mcp/index.js";
 export { listBlocks, lookupBlock } from "./operations/index.js";
