+++ packages/cli/pragma/src/domains/config/index.ts
@@ -12,6 +12,7 @@ export function commands(ctx: PragmaContext): CommandDefinition[] {
   return [tierCommand(ctx), channelCommand(ctx), showCommand(ctx)];
 }
 
+export { specs as mcpSpecs } from "./mcp/index.js";
 export type { ConfigShowData } from "./operations/index.js";
 export {
   resolveConfigShow,
