+++ packages/cli/pragma/src/domains/create/index.ts
@@ -11,3 +11,5 @@ import { componentCommand, packageCommand } from "./commands/index.js";
 export function commands(): CommandDefinition[] {
   return [componentCommand(), packageCommand()];
 }
+
+export { specs as mcpSpecs } from "./mcp/index.js";
