+++ packages/cli/pragma/src/domains/doctor/index.ts
@@ -1,4 +1,5 @@
 export { doctorCommand } from "./commands/index.js";
+export { specs as mcpSpecs } from "./mcp/index.js";
 export type {
   CheckContext,
   CheckResult,
