+++ packages/cli/pragma/src/domains/graph/index.ts
@@ -12,4 +12,5 @@ export function commands(ctx: PragmaContext): CommandDefinition[] {
   return [queryCommand(ctx), inspectCommand(ctx)];
 }
 
+export { specs as mcpSpecs } from "./mcp/index.js";
 export { executeQuery, inspectUri } from "./operations/index.js";
