+++ packages/cli/pragma/src/domains/llm/index.ts
@@ -1,1 +1,2 @@
 export { buildLlmCommand } from "./commands/index.js";
+export { specs as mcpSpecs } from "./mcp/index.js";
