+++ packages/cli/pragma/src/domains/ontology/index.ts
@@ -12,6 +12,7 @@ export function commands(ctx: PragmaContext): CommandDefinition[] {
   return [listCommand(ctx), showCommand(ctx)];
 }
 
+export { specs as mcpSpecs } from "./mcp/index.js";
 export {
   listOntologies,
   showOntology,
