+++ packages/cli/pragma/src/domains/shared/filters/describeFilters.ts
@@ -1,0 +1,14 @@
+/**
+ * Describe active filters as human-readable key-value pairs.
+ */
+
+import type { FilterConfig } from "../types.js";
+
+export default function describeFilters(
+  filters: FilterConfig,
+): Record<string, string> {
+  return {
+    ...(filters.tier !== undefined && { tier: filters.tier }),
+    channel: filters.channel,
+  };
+}
