+++ packages/cli/pragma/src/domains/shared/filters/index.ts
@@ -1,7 +1,9 @@
 export { buildChannelFilter, CHANNEL_RELEASES } from "./buildChannelFilter.js";
+export { default as buildFilterConfig } from "./buildFilterConfig.js";
 export { buildFilters } from "./buildFilters.js";
 export {
   buildTierFilter,
   resolveTierChain,
   tierPathToLocal,
 } from "./buildTierFilter.js";
+export { default as describeFilters } from "./describeFilters.js";
