+++ packages/cli/pragma/src/domains/skill/index.ts
@@ -12,5 +12,6 @@ export function commands(ctx: PragmaContext): CommandDefinition[] {
   return [listCommand(ctx)];
 }
 
+export { specs as mcpSpecs } from "./mcp/index.js";
 export type { SkillListResult } from "./operations/index.js";
 export { discoverSkills, listSkills } from "./operations/index.js";
