+++ packages/cli/pragma/src/domains/standard/formatters/index.ts
@@ -1,3 +1,4 @@
 export { default as categoriesFormatters } from "./categories.js";
 export { default as listFormatters } from "./list.js";
 export { default as lookupFormatters } from "./lookup.js";
+export type { StandardListOutput, StandardLookupInput } from "./types.js";
