+++ packages/cli/pragma/src/domains/standard/index.ts
@@ -17,6 +17,7 @@ export function commands(ctx: PragmaContext): CommandDefinition[] {
   return [listCommand(ctx), lookupCommand(ctx), categoriesCommand(ctx)];
 }
 
+export { specs as mcpSpecs } from "./mcp/index.js";
 export {
   listCategories,
   listStandards,
