+++ packages/cli/pragma/src/domains/standard/mcp/index.ts
@@ -1,0 +1,1 @@
+export { default as specs } from "./specs.js";
