+++ packages/cli/pragma/src/domains/tier/index.ts
@@ -6,4 +6,5 @@ export function commands(ctx: PragmaContext): CommandDefinition[] {
   return [listCommand(ctx)];
 }
 
+export { specs as mcpSpecs } from "./mcp/index.js";
 export { listTiers } from "./operations/index.js";
