+++ packages/cli/pragma/src/domains/token/index.ts
@@ -10,6 +10,7 @@ export function commands(ctx: PragmaContext): CommandDefinition[] {
   return [listCommand(ctx), lookupCommand(ctx), addConfigCommand()];
 }
 
+export { specs as mcpSpecs } from "./mcp/index.js";
 export type { AddConfigResult, TokenListFilters } from "./operations/index.js";
 export {
   listTokens,
